+++ src/components/Header.ts
@@ -1,4 +1,4 @@
-import { html } from "../utils";
+import { html } from "@/lib/utils";
 import Component from "./core/Component";
 
 export default class Header extends Component {
