+++ src/components/core/index.ts
@@ -1,0 +1,1 @@
+export { default as Component } from './Component';
