+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export { default as Footer } from "./Footer";
+export { default as Header } from "./Header";
+export { default as MovieListSkeleton } from "./MovieListSkeleton";
+export { default as Movies } from "./Movies";
+export { default as Skeleton } from "./Skeleton";
+export { default as ThumbnailList } from "./ThumbnailList";
